call ref.reinterpret() when getting/setting from a Buffer that is too small
This is necessary for when a "ref-array" is a return value to an FFI'd function,
where the underlying buffer is 0 bytes. This may or may not be a band-aid for
now, but the idea is that you manually set the ".length" property of the
returned array instance before working with it.